+++ libcaf_core/caf/actor_pool.cpp
@@ -118,14 +118,6 @@ const char* actor_pool::name() const {
   return "caf.actor-pool";
 }
 
-void actor_pool::on_destroy() {
-  CAF_PUSH_AID_FROM_PTR(this);
-  if (!getf(is_cleaned_up_flag)) {
-    cleanup(exit_reason::unreachable, nullptr);
-    unregister_from_system();
-  }
-}
-
 void actor_pool::on_cleanup(const error& reason) {
   CAF_PUSH_AID_FROM_PTR(this);
   CAF_IGNORE_UNUSED(reason);
@@ -220,6 +212,10 @@ void actor_pool::quit(execution_unit* host) {
   // workers_mtx_ because abstract_actor has its own lock
   if (cleanup(planned_reason_, host))
     unregister_from_system();
+}
+
+void actor_pool::force_close_mailbox() {
+  // nop
 }
 
 } // namespace caf