+++ libcaf_core/caf/actor_pool.hpp
@@ -90,8 +90,6 @@ public:
 
   const char* name() const override;
 
-  void on_destroy() override;
-
   void setup_metrics() {
     // nop
   }
@@ -105,6 +103,8 @@ private:
 
   // call without workers_mtx_ held
   void quit(execution_unit* host);
+
+  void force_close_mailbox() override;
 
   std::mutex workers_mtx_;
   std::vector<actor> workers_;