+++ libcaf_net/caf/net/abstract_actor_shell.cpp
@@ -160,16 +160,9 @@ void abstract_actor_shell::launch(execution_unit*, bool, bool hide) {
     register_at_system();
 }
 
-bool abstract_actor_shell::cleanup(error&& fail_state, execution_unit* host) {
-  CAF_LOG_TRACE(CAF_ARG(fail_state));
-  // Clear mailbox.
-  if (!mailbox_.closed()) {
-    auto dropped = mailbox_.close(fail_state);
-    while (dropped > 0 && getf(abstract_actor::collects_metrics_flag)) {
-      auto val = static_cast<int64_t>(dropped);
-      metrics_.mailbox_size->dec(val);
-    }
-  }
+void abstract_actor_shell::on_cleanup(const error& reason) {
+  CAF_LOG_TRACE(CAF_ARG(reason));
+  close_mailbox(reason);
   // Detach from owner.
   {
     std::unique_lock<std::mutex> guard{loop_mtx_};
@@ -177,7 +170,21 @@ bool abstract_actor_shell::cleanup(error&& fail_state, execution_unit* host) {
     resume_.dispose();
   }
   // Dispatch to parent's `cleanup` function.
-  return super::cleanup(std::move(fail_state), host);
+  return super::on_cleanup(reason);
+}
+
+// -- private member functions -------------------------------------------------
+
+void abstract_actor_shell::close_mailbox(const error& reason) {
+  if (!mailbox_.closed()) {
+    auto dropped = mailbox_.close(reason);
+    if (dropped > 0 && metrics_.mailbox_size)
+      metrics_.mailbox_size->dec(static_cast<int64_t>(dropped));
+  }
+}
+
+void abstract_actor_shell::force_close_mailbox() {
+  close_mailbox(make_error(exit_reason::unreachable));
 }
 
 } // namespace caf::net