+++ libcaf_net/caf/net/abstract_actor_shell.hpp
@@ -98,7 +98,7 @@ public:
 
   void launch(execution_unit* eu, bool lazy, bool hide) override;
 
-  bool cleanup(error&& fail_state, execution_unit* host) override;
+  void on_cleanup(const error& reason) override;
 
 protected:
   void set_behavior_impl(behavior bhvr) {
@@ -106,6 +106,10 @@ protected:
   }
 
 private:
+  void close_mailbox(const error& reason);
+
+  void force_close_mailbox() final;
+
   /// Stores incoming actor messages.
   detail::default_mailbox mailbox_;